refactor: extract file upload modal as a separate component
