+++ app/views/Data/Upload/UploadModal/styles.module.css
@@ -1,0 +1,3 @@
+.paper {
+    cursor: pointer;
+}
