+++ app/views/Data/Upload/styles.module.css
@@ -1,4 +1,4 @@
-.drop-zone {
+.import {
     display: flex;
     flex-direction: column;
     gap: var(--mantine-spacing-md);
@@ -7,8 +7,4 @@
     border-radius: var(--mantine-radius-md);
     padding: var(--mantine-spacing-md);
     max-width: 250px;
-}
-
-.paper {
-    cursor: pointer;
 }
