+++ schema.graphql
@@ -183,17 +183,8 @@ schema (one of the key benefits of GraphQL).
 """
 scalar JSONString
 
-type KeyValueType {
-  key: String
-  label: String
-}
-
 type Mutation {
   createDataset(data: CreateDatasetInputType!): CreateDataset
-}
-
-type PropertiesType {
-  table: TablePropertiesType
 }
 
 type Query {
@@ -236,13 +227,10 @@ type Query {
     """Number of results to return per page. Default 'page_size': 20"""
     pageSize: Int
   ): FileListType
-  properties: PropertiesType
-}
-
-type TablePropertiesType {
-  headers: [KeyValueType]
-  languages: [KeyValueType]
-  timeZones: [KeyValueType]
+  table(
+    """Django object unique identification field"""
+    id: ID!
+  ): TableType
 }
 
 """An enumeration."""
@@ -263,6 +251,7 @@ type TableType {
 
   """"""
   name: String!
+  previewData: GenericScalar
 
   """"""
   status: TableStatus!
